added progress indicator and moved service mock module
