+++ Fabric.Authorization.AccessControl/src/app/login/login.component.html
@@ -1,6 +1,11 @@
 <div *ngIf="loggingIn">
   <label>Logging in</label>
-  <i class="fa fa-spinner fa-spin fa-3x fa-fw"></i>
+  <hc-progress-spinner [diameter]="60"
+    [color]="'blue'"
+    [hasChannel]="true"
+    [isCentered]="false"
+    [isDeterminate]="false">
+    </hc-progress-spinner>
   <span class="sr-only">Loading...</span>
 </div>
 <h1 [innerText]="status"></h1>
