+++ Fabric.Authorization.AccessControl/src/app/logout/logout.component.html
@@ -1,5 +1,10 @@
 <div>
     <label>Logging out</label>
-    <i class="fa fa-spinner fa-spin fa-3x fa-fw"></i>
+    <hc-progress-spinner [diameter]="60"
+    [color]="'blue'"
+    [hasChannel]="true"
+    [isCentered]="false"
+    [isDeterminate]="false">
+    </hc-progress-spinner>
     <span class="sr-only">Loading...</span>
 </div>
