+++ Fabric.Authorization.AccessControl/src/app/modules/access-control/member-edit/member-edit.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed, inject } from '@angular/core/testing';
 
 import { MemberEditComponent } from './member-edit.component';
-import { ServicesMockModule } from '../member-add/services.mock.module';
+import { ServicesMockModule } from '../services.mock.module';
 import { FabricAuthGroupService, FabricAuthRoleService } from '../../../services';
 import { FabricAuthGroupServiceMock } from '../../../services/fabric-auth-group.service.mock';
 import { Observable } from 'rxjs/Observable';
